# integrated-api/src/main/java/com/backend/integratedapi/indexingjob/controller/IndexingJobController.java
package com.backend.integratedapi.indexingjob.controller;

import com.backend.commondataaccess.dto.OffsetPageResult;
import com.backend.commondataaccess.persistence.common.enums.JobStatus;
import com.backend.commondataaccess.security.CurrentUser;
import com.backend.commondataaccess.security.JwtPrincipal;
import com.backend.integratedapi.indexingjob.controller.dto.IndexingJobReadDto;
import com.backend.integratedapi.indexingjob.controller.dto.IndexingJobStartDto;
import com.backend.integratedapi.indexingjob.service.IndexingJobService;
import com.backend.integratedapi.indexingjob.service.dto.IndexingJobDto;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import java.util.UUID;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

@Tag(name = "05. 색인 작업 실행 API")
@RequestMapping(value = "/index/v1", produces = MediaType.APPLICATION_JSON_VALUE)
@RestController
@RequiredArgsConstructor
public class IndexingJobController {

    private final IndexingJobService indexingJobService;

    @Operation(summary = "이 '수집 소스'에 포함된 전체 post 재색인", description = "운영자 전용 API")
    @ResponseStatus(HttpStatus.ACCEPTED)
    @PostMapping("/sources/{source-id}/_reindex")
    public ResponseEntity<IndexingJobStartDto.Response> reindexSource(@CurrentUser JwtPrincipal principal,
                                                                      @PathVariable("source-id") UUID sourceId) {

        UUID jobId = indexingJobService.triggerReindexByCollectSource(sourceId, principal.getUserId()).id();
        return ResponseEntity.accepted().body(new IndexingJobStartDto.Response(jobId, JobStatus.PENDING));
    }

    @Operation(summary = "특정 post 1개 재색인", description = "운영자 전용 API")
    @ResponseStatus(HttpStatus.ACCEPTED)
    @PostMapping("/posts/{post-id}/_reindex")
    public ResponseEntity<IndexingJobStartDto.Response> reindexPost(@CurrentUser JwtPrincipal principal,
                                                                    @PathVariable("post-id") UUID postId) {

        UUID jobId = indexingJobService.triggerReindexByCollectSourcePost(postId, principal.getUserId()).id();
        return ResponseEntity.accepted().body(new IndexingJobStartDto.Response(jobId, JobStatus.PENDING));
    }

    @Operation(summary = "색인 작업 상태 목록 조회", description = "운영자 전용 API")
    @GetMapping("/jobs")
    public ResponseEntity<OffsetPageResult<IndexingJobReadDto.Response>> getIndexingJobs(@RequestParam(value = "page", required = false, defaultValue = "0") int page,
                                                                                         @RequestParam(value = "size", required = false, defaultValue = "10") int size) {

        OffsetPageResult<IndexingJobDto> indexingJobDtos = indexingJobService.getIndexingJobs(page, size);
        return ResponseEntity.ok(indexingJobDtos.map(IndexingJobReadDto.Response::from));
    }

    @Operation(summary = "색인 작업 상태 조회", description = "운영자 전용 API")
    @GetMapping("/jobs/{id}")
    public ResponseEntity<IndexingJobReadDto.Response> getIndexingJob(@PathVariable UUID id) {

        IndexingJobDto indexingJobDto = indexingJobService.getIndexingJobDto(id);
        return ResponseEntity.ok(IndexingJobReadDto.Response.from(indexingJobDto));
    }
}